Refactoring variance handling
